+++ lib/mssql.dart
@@ -16,3 +16,4 @@ export 'src/server_endpoint.dart';
 export 'src/tds/bulk.dart';
 export 'src/tds/sql_browser.dart';
 export 'src/tds/tvp.dart';
+export 'src/typed_values.dart';